Implement single function Turnstile.
Remove the event emitter and all that non-sense about invocation with
parameters and whatnot. The guarded action is encapsulated in a closure
that accepts no parameters. Entering the Turnstile returns nothing.
There is no way to wait for the Turnstile to complete. You send a
message and go on with your life.

Resuming on the other side presents and error handling problem along the
lines of a server handling a request. There is no good place to report a
fatal error. This is nothing that should be solved in Turnstile itself.

Closes #15.