Fix abnormal exits.
Updated `abend` so that the exception would always be raised where it
cannot be caught; in the next tick. Created `eject` to do what `abend`
did. Testing is complete.

After use in an application, it will make sense to create three modules
in this project, `turnstile.js`, `throttle.js` and `buffer.js`.

Considering how to create a queue that operates on intervals.