Convert `Turnstile` to ES6.
Convert `Turnstile` to ES6 with a new interface, of course, but really
new since `Fracture` is not implemented as a collection of `Turnstile`s
so there is no need for an exit notice, and I could have wrapped the
function anyway, but I probably didn't want to create anonymous
error-first callback functions. Creating anonymous async functions is
not a big deal.

Convert the `Turnstile` test to Mocha.

Closes #138.
See #139.