Survive errors, Destructible shutdown.
Now surviving errors, running through the existing queue, then reporting
the exception with a Destructible with nested errors.

Also, merging an object argument with the entry so that the callback
function can use a single level of destructuring to receive the
arguments of interest.

Removed the rejector queue. There was a bug in it that would cause the
`waiting` count to be wrong and the primary strands would run past the
end of the queue. Repurposed the primary worker function to optionally
operate as a rejector strand.