Create countdown object, test `unqueue`.
Renamed `dequeue` to `unqueue` and wrote a test for it.

Created a countdown object that contains a `destructible` that is
durable and will be destroyed with the associated Turnstile gets
destroyed. Users can register destructors on this object to flush their
work before decrementing the countdown. Also contains a reference to the
turnstile. This object can be passed into constructors of object that
depend on the turnstile.