Create a Countdown object.
Users need a way to hold the Turnstile open while they finish up their
processing. We used to say that once the Turnstile closes you cannot
submit work, if you want to hold it open you can manage Destructible
countdowns yourself, except that now we find that there's almost never a
case where countdown is not necessary. May as well make it a feature of
Turnstile.

We expose the Destructible-ish-ness of this in the countdown, creating a
durable destructible that the user can use to get a destruct
notification and to launch an ephemeral for their own drain.