Use Destructible.Counter.
Counter is now a convention and an abstraction and Turnstile is the
first to implement this convention. It exposes a `destructible` that is
a terminal destructible. If you destory it it will shutdown the
destructible that created the Turnstile. Not sure if this is correct,
acutally. Anyway, `countdown` is where the real work is done. `counted`
is an object that contains the counted object.