Complete shutdown on error.
Now that we have isolation we can isolate the Turnstile so that it
doesn't catch an exception and have Fracture run an applications workers
in a `destructible.destructive()` call that can catch the exception and
destroy the Fracture. Fracture can then blow through it's work queue
rejecting all the completed futures or call a user provided error
cleanup function.